Refactor Snooty Preview functions into separate file
